+++ src/components/Feed.jsx
@@ -3,8 +3,8 @@ import React, { useState } from 'react';
 import { Text, Image, TouchableOpacity, View, ScrollView } from 'react-native';
 import { Link } from '@react-navigation/native';
 import { icons } from '../constants/icons';
-import Video from 'react-native-video';
 import CustomVideoPlayer from './CustomVideoPlayer';
+import { useNavigation } from '@react-navigation/native';
 
 const formatNumber = (num) => {
     if (num >= 1_000_000) return (num / 1_000_000).toFixed(1) + 'M';
@@ -29,16 +29,30 @@ const formatDate = (date) => {
     if (diffDays <= 7) return `${diffDays} ngày trước`;
     return givenDate.toLocaleDateString('vi-VN'); // Hiển thị ngày tháng đầy đủ
 };
+
 const Feed = ( {thread} ) => {
     const [like, setLike] = useState(false);
+    const navigation = useNavigation();
     
+    const handleGoProfile = (id) => {
+        navigation.navigate("UserProfile", { id })
+    }
+
+    const handleReply = (id) => {
+        navigation.navigate('FeedDetail', { id });
+    };
+
     return(
         <View className='flex-row items-center px-3 py-4 gap-1'>
-            <Image source={{ uri: thread.avatar_path }} className='w-14 h-14 rounded-full self-start'/> 
+            <TouchableOpacity onPress={() => handleGoProfile(thread.authorId)} className='self-start'>
+                <Image source={{ uri: thread.avatar_path }} className='w-14 h-14 rounded-full'/> 
+            </TouchableOpacity>
         <View className='flex-1 gap-1 ml-4' >
             <View className='flex-row items-center'>
                 <View className='flex-row items-center flex-1 gap-1.5'>
-                    <Text className='text-base font-bold' numberOfLines={1} style={{ flexShrink: 1 }}>{thread.fullname}</Text>
+                    <TouchableOpacity onPress={() => handleGoProfile(thread.authorId)}>
+                        <Text className='text-base font-bold' numberOfLines={1} style={{ flexShrink: 1 }}>{thread.fullname}</Text>
+                    </TouchableOpacity>
                     <Text className='text-sm text-gray-500' >{formatDate(thread.createdAt)}</Text>    
                 </View>
                 <Image source={icons.more} className='size-6 self-end' tintColor="gray"/>
@@ -85,7 +99,7 @@ const Feed = ( {thread} ) => {
                             </>
                         )}
                 </TouchableOpacity>
-                <TouchableOpacity className="flex-row items-center">
+                <TouchableOpacity className="flex-row items-center" onPress={() => handleReply(thread.threadid)}>
                         <Image source={icons.chat} className='size-6' />
                         <Text className='text-base font-normal ml-1' >{formatNumber(thread.commentCount)}</Text>
                 </TouchableOpacity>
