+++ src/services/threadService.js
@@ -93,7 +93,7 @@ export const getThread = async () => {
     if (snapshot.exists()) {
       const data = snapshot.val();
       const threads = Object.entries(data).map(([key, value]) => ({
-        threadid: key, // ✅ gán key làm threadid
+        threadid: key, //  gán key làm threadid
         ...value,
       }));
       return threads.sort((a, b) => new Date(b.createdAt) - new Date(a.createdAt)); // Sắp xếp theo createdAt giảm dần
